Split training and interp into their own files
