+++ src/interventions/auto_finetune_eval/README.md
@@ -7,9 +7,11 @@ This project provides an automated framework for evaluating the effectiveness of
 The project consists of the following main components:
 
 1. `auto_finetuning_data.py`: Generates ground truths and associated datasets using an assistant LLM API, saving them to a CSV file.
-2. `auto_finetuning_compare_to_truth.py`: Compares discovered differences with ground truth differences using LLMs and scores the interpretability method's performance.
-3. `auto_finetuning_main.py`: Orchestrates the entire process, including data generation, model fine-tuning, applying the interpretability method, and evaluation.
-4. `auto_finetuning_helpers.py`: Contains helper functions and dummy implementations for development purposes.
+2. `auto_finetuning_train.py`: Finetunes models using the data generated by `auto_finetuning_data.py`.
+3. `auto_finetuning_interp.py`: Applies the interpretability method to automatically discover the differences between the original and finetuned models.
+4. `auto_finetuning_compare_to_truth.py`: Compares discovered differences with ground truth differences using LLMs and scores the interpretability method's performance.
+5. `auto_finetuning_main.py`: Orchestrates the entire process, including data generation, model fine-tuning, applying the interpretability method, and evaluation.
+6. `auto_finetuning_helpers.py`: Contains helper functions for making API requests, loading API keys and parsing strings for json.
 
 ## Key Features
