+++ src/interventions/auto_finetune_eval/auto_finetuning_train.py
@@ -1,0 +1,90 @@
+"""
+This module contains code for finetuning models on the data associated with multiple ground truths.
+"""
+
+from typing import List, Dict, Any, Optional
+from transformers import PreTrainedModel, PreTrainedTokenizer, TrainingArguments, Trainer, DataCollatorForLanguageModeling
+from datasets import Dataset
+#import torch
+#from bitsandbytes import BitsAndBytesConfig
+
+
+def dummy_finetune_model(
+    base_model: PreTrainedModel,
+    tokenizer: PreTrainedTokenizer,
+    training_data: List[str],
+    finetuning_params: Dict[str, Any]
+) -> PreTrainedModel:
+    """
+    Dummy implementation of finetuning a model on given training data.
+
+    This function simulates the process of finetuning a model. In a real implementation,
+    this would involve actual training on the provided data using the specified parameters.
+
+    Args:
+        base_model (PreTrainedModel): The original model to be finetuned.
+        tokenizer (PreTrainedTokenizer): The tokenizer associated with the model.
+        training_data (List[Dict[str, str]]): List of training examples, each a dict with 'input' and 'output' keys.
+        finetuning_params (Dict[str, Any]): Parameters for the finetuning process.
+
+    Returns:
+        PreTrainedModel: A "finetuned" version of the input model (in this dummy implementation, it's the same model).
+    """
+    # Placeholder implementation
+    print(f"Finetuning model with {len(training_data)} examples and parameters: {finetuning_params}")
+    return base_model  # In reality, this would be a new, finetuned model
+
+def finetune_model(
+        base_model: PreTrainedModel, 
+        tokenizer: PreTrainedTokenizer, 
+        training_data: List[str], 
+        finetuning_params: Dict[str, Any]
+    ) -> PreTrainedModel:
+    """
+    Fine-tune a pre-trained model on the given training data.
+
+    Args:
+        base_model (PreTrainedModel): The original model to be fine-tuned.
+        tokenizer (PreTrainedTokenizer): The tokenizer associated with the model.
+        training_data (List[str]): List of training examples.
+        finetuning_params (Dict[str, Any]): Parameters for the fine-tuning process.
+
+    Returns:
+        PreTrainedModel: The fine-tuned model.
+    """
+    # Prepare the dataset
+    def tokenize_function(examples):
+        return tokenizer(examples["text"], padding="max_length", truncation=True)
+
+    dataset = Dataset.from_dict({"text": training_data})
+    tokenized_dataset = dataset.map(tokenize_function, batched=True)
+
+    # Set up the data collator
+    data_collator = DataCollatorForLanguageModeling(tokenizer=tokenizer, mlm=False)
+
+    # Set up training arguments
+    training_args = TrainingArguments(
+        output_dir="./results",
+        num_train_epochs=finetuning_params.get("num_epochs", 3),
+        per_device_train_batch_size=finetuning_params.get("batch_size", 8),
+        warmup_steps=finetuning_params.get("warmup_steps", 500),
+        weight_decay=finetuning_params.get("weight_decay", 0.01),
+        logging_dir="./logs",
+        logging_steps=finetuning_params.get("logging_steps", 100),
+        save_steps=finetuning_params.get("save_steps", 1000),
+        learning_rate=finetuning_params.get("learning_rate", 5e-5),
+    )
+
+    # Initialize the Trainer
+    trainer = Trainer(
+        model=base_model,
+        args=training_args,
+        train_dataset=tokenized_dataset,
+        data_collator=data_collator,
+    )
+
+    # Fine-tune the model
+    trainer.train()
+
+    # Return the fine-tuned model
+    return trainer.model
